refactor: move embeddings to its own module and seperate embeddable
