+++ rig-core/rig-core-derive/src/lib.rs
@@ -10,7 +10,7 @@ pub(crate) const EMBED: &str = "embed";
 // https://doc.rust-lang.org/book/ch19-06-macros.html#how-to-write-a-custom-derive-macro
 // https://doc.rust-lang.org/reference/procedural-macros.html
 
-#[proc_macro_derive(Embed, attributes(embed))]
+#[proc_macro_derive(Embeddable, attributes(embed))]
 pub fn derive_embedding_trait(item: TokenStream) -> TokenStream {
     let mut input = parse_macro_input!(item as DeriveInput);
