+++ rig-core/src/embeddings/embeddable.rs
@@ -1,20 +1,6 @@
-//! This module provides functionality for working with embeddings and embedding models.
-//! Embeddings are numerical representations of documents or other objects, typically used in
-//! natural language processing (NLP) tasks such as text classification, information retrieval,
-//! and document similarity.
-//!
-//! The module defines the [EmbeddingModel] trait, which represents an embedding model that can
-//! generate embeddings for documents. It also provides an implementation of the [EmbeddingsBuilder]
-//! struct, which allows users to build collections of document embeddings using different embedding
-//! models and document sources.
-//!
-//! The module also defines the [Embedding] struct, which represents a single document embedding.
-//!
-//! The module also defines the [Embeddable] trait, which represents types that can be embedded.
-//! Only types that implement the Embeddable trait can be used with the EmbeddingsBuilder.
-//!
-//! Finally, the module defines the [EmbeddingError] enum, which represents various errors that
-//! can occur during embedding generation or processing.
+//! The module defines the [Embeddable] trait, which must be implemented for types that can be embedded.
+//! The module defines the [EmbeddingsBuilder] struct which accumulates objects to be embedded and generates the embeddings for each object when built.
+//! Only types that implement the [Embeddable] trait can be added to the [EmbeddingsBuilder].
 //!
 //! # Example
 //! ```rust
@@ -73,102 +59,18 @@
 //! // ...
 //! ```
 
+use super::embedding::{Embedding, EmbeddingError, EmbeddingModel};
+use futures::{stream, StreamExt, TryStreamExt};
 use std::{cmp::max, collections::HashMap, marker::PhantomData};
 
-use futures::{stream, StreamExt, TryStreamExt};
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, thiserror::Error)]
-pub enum EmbeddingError {
-    /// Http error (e.g.: connection error, timeout, etc.)
-    #[error("HttpError: {0}")]
-    HttpError(#[from] reqwest::Error),
-
-    /// Json error (e.g.: serialization, deserialization)
-    #[error("JsonError: {0}")]
-    JsonError(#[from] serde_json::Error),
-
-    /// Error processing the document for embedding
-    #[error("DocumentError: {0}")]
-    DocumentError(String),
-
-    /// Error parsing the completion response
-    #[error("ResponseError: {0}")]
-    ResponseError(String),
-
-    /// Error returned by the embedding model provider
-    #[error("ProviderError: {0}")]
-    ProviderError(String),
-}
-
-/// Trait for embedding models that can generate embeddings for documents.
-pub trait EmbeddingModel: Clone + Sync + Send {
-    /// The maximum number of documents that can be embedded in a single request.
-    const MAX_DOCUMENTS: usize;
-
-    /// The number of dimensions in the embedding vector.
-    fn ndims(&self) -> usize;
-
-    /// Embed a single document
-    fn embed_document(
-        &self,
-        document: &str,
-    ) -> impl std::future::Future<Output = Result<Embedding, EmbeddingError>> + Send
-    where
-        Self: Sync,
-    {
-        async {
-            Ok(self
-                .embed_documents(vec![document.to_string()])
-                .await?
-                .first()
-                .cloned()
-                .expect("One embedding should be present"))
-        }
-    }
-
-    /// Embed multiple documents in a single request
-    fn embed_documents(
-        &self,
-        documents: Vec<String>,
-    ) -> impl std::future::Future<Output = Result<Vec<Embedding>, EmbeddingError>> + Send;
-}
-
-/// Struct that holds a single document and its embedding.
-#[derive(Clone, Default, Deserialize, Serialize, Debug)]
-pub struct Embedding {
-    /// The document that was embedded. Used for debugging.
-    pub document: String,
-    /// The embedding vector
-    pub vec: Vec<f64>,
-}
-
-impl PartialEq for Embedding {
-    fn eq(&self, other: &Self) -> bool {
-        self.document == other.document
-    }
-}
-
-impl Eq for Embedding {}
-
-impl Embedding {
-    pub fn distance(&self, other: &Self) -> f64 {
-        let dot_product: f64 = self
-            .vec
-            .iter()
-            .zip(other.vec.iter())
-            .map(|(x, y)| x * y)
-            .sum();
-
-        let product_of_lengths = (self.vec.len() * other.vec.len()) as f64;
-
-        dot_product / product_of_lengths
-    }
-}
-
+/// The associated type `Kind` on the trait `Embeddable` must implement this trait.
 pub trait EmbeddingKind {}
+
+/// Used for structs that contain a single embedding target.
 pub struct SingleEmbedding;
 impl EmbeddingKind for SingleEmbedding {}
+
+/// Used for structs that contain many embedding targets.
 pub struct ManyEmbedding;
 impl EmbeddingKind for ManyEmbedding {}
 
@@ -178,6 +80,9 @@ pub enum EmbeddingGenerationError {
     SerdeError(#[from] serde_json::Error),
 }
 
+/// Trait for types that can be embedded.
+/// The `embeddable` method returns a list of strings for which embeddings will be generated by the embeddings builder.
+/// If the type `Kind` is `SingleEmbedding`, the list of strings contains a single item, otherwise, the list can contain many items.
 pub trait Embeddable {
     type Kind: EmbeddingKind;
     type Error: std::error::Error;
@@ -185,7 +90,7 @@ pub trait Embeddable {
     fn embeddable(&self) -> Result<Vec<String>, Self::Error>;
 }
 
-/// Builder for creating a collection of embeddings
+/// Builder for creating a collection of embeddings.
 pub struct EmbeddingsBuilder<M: EmbeddingModel, D: Embeddable, K: EmbeddingKind> {
     kind: PhantomData<K>,
     model: M,
@@ -431,11 +336,10 @@ impl<T: Embeddable> Embeddable for Vec<T> {
 
 #[cfg(test)]
 mod tests {
-    use super::{Embeddable, SingleEmbedding, EmbeddingGenerationError};
+    use super::{Embeddable, EmbeddingGenerationError, SingleEmbedding};
+    use rig_derive::Embeddable;
 
-    use rig_derive::Embed;
-
-    #[derive(Embed)]
+    #[derive(Embeddable)]
     struct FakeDefinition {
         id: String,
         #[embed]
