+++ rig-core/src/embeddings/mod.rs
@@ -1,0 +1,7 @@
+//! This module provides functionality for working with embeddings.
+//! Embeddings are numerical representations of documents or other objects, typically used in
+//! natural language processing (NLP) tasks such as text classification, information retrieval,
+//! and document similarity.
+
+pub mod embeddable;
+pub mod embedding;
