+++ rig-core/src/vector_store/in_memory_store.rs
@@ -8,7 +8,7 @@ use ordered_float::OrderedFloat;
 use serde::{Deserialize, Serialize};
 
 use super::{VectorStoreError, VectorStoreIndex};
-use crate::embeddings::{Embedding, EmbeddingModel};
+use crate::embeddings::embedding::{Embedding, EmbeddingModel};
 
 /// InMemoryVectorStore is a simple in-memory vector store that stores embeddings
 /// in-memory using a HashMap.
